requirements.txt
Pillow~=9.1.0
Flask==2.1.1
ISR==2.2.0
gunicorn
requests[security]
